third_party/WebKit/LayoutTests/http/tests/htmlimports/import-and-script-module-inline.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="import" href="resources/module-script-inline.html">
<script>
test(() => {
  assert_true(window.runInlineScript);
}, "Inline module script ran in HTML import.");
</script>
